用户相关业务完成

+++ spzx-manager/src/main/java/com/startzhao/spzx/manager/service/impl/SysRoleServiceImpl.java
@@ -7,9 +7,11 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.startzhao.spzx.common.exception.StartZhaoException;
 import com.startzhao.spzx.manager.mapper.SysRoleMapper;
+import com.startzhao.spzx.manager.mapper.SysRoleUserMapper;
 import com.startzhao.spzx.manager.service.SysRoleService;
 import com.startzhao.spzx.model.dto.system.SysRoleDTO;
 import com.startzhao.spzx.model.entity.system.SysRole;
+import com.startzhao.spzx.model.entity.system.SysRoleUser;
 import com.startzhao.spzx.model.vo.common.PageResult;
 import com.startzhao.spzx.model.vo.common.Result;
 import com.startzhao.spzx.model.vo.common.ResultCodeEnum;
@@ -18,7 +20,10 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * ClassName: SysRoleServiceImpl
@@ -35,6 +40,9 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
 
     @Autowired
     private SysRoleMapper sysRoleMapper;
+
+    @Autowired
+    private SysRoleUserMapper sysRoleUserMapper;
 
 
     /**
@@ -87,7 +95,7 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
         String roleName = sysRole.getRoleName();
         String roleCode = sysRole.getRoleCode();
         if (StringUtils.isEmpty(roleName) || StringUtils.isEmpty(roleCode)) {
-            throw new StartZhaoException(500,"角色名或角色编码为空");
+            throw new StartZhaoException(500, "角色名或角色编码为空");
         }
         sysRole.setUpdateTime(DateTime.now());
         sysRoleMapper.updateById(sysRole);
@@ -96,13 +104,45 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
 
     /**
      * 删除角色信息
+     * 由于存在用户与角色信息相关联的情况
+     * 故要在角色与用户没有关联时才可删除
      *
      * @param roleId
      * @return
      */
     @Override
     public Result deleteSysRole(Long roleId) {
+
+        QueryWrapper<SysRoleUser> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("role_id", roleId).eq("is_deleted", 0);
+        List<SysRoleUser> sysRoleUsers = sysRoleUserMapper.selectList(queryWrapper);
+        if (sysRoleUsers.size() > 0) throw new StartZhaoException(500, "该角色被用户关联，若要删除该角色，请先取消用户与角色关联关系");
         sysRoleMapper.deleteById(roleId);
-        return Result.build(null,ResultCodeEnum.SUCCESS);
+        return Result.build(null, ResultCodeEnum.SUCCESS);
+    }
+
+    /**
+     * 1、查询所有角色
+     * 2、根据用户id得到用户对应角色id
+     *
+     * @param userId
+     * @return
+     */
+    @Override
+    public Result<Map<String, List>> findAllRoles(Long userId) {
+        Map<String, List> data = new HashMap<>();
+        QueryWrapper<SysRole> wrapper = new QueryWrapper<>();
+        wrapper.eq("is_deleted",0);
+        List<SysRole> sysRoles = sysRoleMapper.selectList(wrapper);
+        data.put("allRoleList", sysRoles);
+        QueryWrapper<SysRoleUser> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("user_id", userId).eq("is_deleted", 0);
+        List<SysRoleUser> sysRoleUsers = sysRoleUserMapper.selectList(queryWrapper);
+        List<Long> roleIds = new ArrayList<>();
+        sysRoleUsers.forEach(sysRoleUser -> {
+            roleIds.add(sysRoleUser.getRoleId());
+        });
+        data.put("userRoleIds", roleIds);
+        return Result.build(data, ResultCodeEnum.SUCCESS);
     }
 }
